+++ packages/eslint-plugin/rules/__tests__/no-setting-ds-tokens.js
@@ -46,7 +46,7 @@ ruleTester.run( 'no-setting-ds-tokens', rule, {
 			],
 		},
 		{
-			code: `<div style={ { '--wpds-font-size-md': '10px', color: 'blue' } } />`,
+			code: `<div style={ { '--wpds-typography-font-size-md': '10px', color: 'blue' } } />`,
 			errors: [
 				{
 					messageId: 'disallowedSet',
@@ -62,7 +62,7 @@ ruleTester.run( 'no-setting-ds-tokens', rule, {
 			],
 		},
 		{
-			code: `function getStyles() { return { '--wpds-font-size-md': '10px' }; }`,
+			code: `function getStyles() { return { '--wpds-typography-font-size-md': '10px' }; }`,
 			errors: [
 				{
 					messageId: 'disallowedSet',