+++ packages/list-reusable-blocks/src/utils/export.ts
@@ -12,20 +12,26 @@ import apiFetch from '@wordpress/api-fetch';
  * Internal dependencies
  */
 import { downloadBlob } from '@wordpress/blob';
+import type { Post, PostType } from './types';
 
 /**
  * Export a reusable block as a JSON file.
  *
- * @param {number} id
+ * @param id - The ID of the reusable block to export
  */
-async function exportReusableBlock( id ) {
-	const postType = await apiFetch( { path: `/wp/v2/types/wp_block` } );
-	const post = await apiFetch( {
+async function exportReusableBlock( id: number ): Promise< void > {
+	const postType = await apiFetch< PostType >( {
+		path: `/wp/v2/types/wp_block`,
+	} );
+
+	const post = await apiFetch< Post >( {
 		path: `/wp/v2/${ postType.rest_base }/${ id }?context=edit`,
 	} );
+
 	const title = post.title.raw;
 	const content = post.content.raw;
 	const syncStatus = post.wp_pattern_sync_status;
+
 	const fileContent = JSON.stringify(
 		{
 			__file: 'wp_block',
@@ -36,6 +42,7 @@ async function exportReusableBlock( id ) {
 		null,
 		2
 	);
+
 	const fileName = kebabCase( title ) + '.json';
 
 	downloadBlob( fileName, fileContent, 'application/json' );