+++ packages/list-reusable-blocks/src/utils/file.ts
@@ -1,14 +1,14 @@
 /**
  * Reads the textual content of the given file.
  *
- * @param {File} file File.
- * @return {Promise<string>}  Content of the file.
+ * @param file - File to read
+ * @return Promise that resolves to the content of the file
  */
-export function readTextFile( file ) {
+export function readTextFile( file: File ): Promise< string > {
 	const reader = new window.FileReader();
 	return new Promise( ( resolve ) => {
 		reader.onload = () => {
-			resolve( reader.result );
+			resolve( reader.result as string );
 		};
 		reader.readAsText( file );
 	} );